Fix: unable to search by type after deleting one.
Corrects problem that creates a bug then a type is deleted. The expense associated with the type would have its type set to null. That would
make it impossible to filter by type when performing a search.
If type is deleted, now the expenses associeated with it will have their types set to "Other"

A new route was created. A put request is made whenever a type is deleted to update the expeses associeate with it.

"Other" type cannot be deleted.